rewrite CardPopup accouring to changes in card props
