+++ app/javascript/components/CardColumn/Card/card-user/card-user.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {getUserInitials} from '../../../../utils/helpers';
+import CardUserAvatar from '../card-user-avatar/card-user-avatar';
 
 const CardUser = ({
   first_name,
@@ -9,21 +9,13 @@ const CardUser = ({
     thumb: {url}
   }
 }) => {
-  const renderAvatar = (userAvatar, userName, userSurname) => {
-    if (userAvatar) {
-      return <img src={userAvatar} className="avatar" />;
-    }
-
-    return (
-      <div className="avatar avatar--text">
-        {getUserInitials(userName, userSurname)}
-      </div>
-    );
-  };
-
   return (
     <div className="column avatar__container">
-      {renderAvatar(url, first_name, last_name)}
+      <CardUserAvatar
+        avatar={url}
+        firstName={first_name}
+        lastName={last_name}
+      />
       <span className="avatar__nickname"> {nickname}</span>
     </div>
   );
